refactor: repository 클래스명 ChessService로 변경
