+++ src/main/java/service/ChessService.java
@@ -1,5 +1,8 @@
-package db;
+package service;
 
+import db.BoardDao;
+import db.MovingDao;
+import db.TurnDao;
 import db.dto.BoardDto;
 import db.dto.MovingDto;
 import db.dto.TurnDto;
@@ -12,13 +15,13 @@ import model.Turn;
 import model.position.Moving;
 import model.position.Position;
 
-public class Repository { // TODO service아님? 왜 repository임
+public class ChessService {
 
     private final MovingDao movingDao;
     private final TurnDao turnDao;
     private final BoardDao boardDao;
 
-    public Repository(final String database) {
+    public ChessService(final String database) {
         this.movingDao = new MovingDao(database);
         this.turnDao = new TurnDao(database);
         this.boardDao = new BoardDao(database);
